+++ src/components/organisms/comment/render/Comment.tsx
@@ -1,24 +1,24 @@
+import { useState } from "react";
+
 import LinkTo from "@Components/atom/LinkTo";
 import PublishDate from "@Components/atom/text/PublishDate";
 import UserProfile from "@Components/molecules/profile/UserProfile";
 import CommentContents from "@Components/organisms/comment/content/CommentContents";
 import CommentUpdateInput from "@Components/organisms/comment/input/CommentUpdateInput";
 import CommentTools from "@Components/organisms/comment/tools/CommentTools";
-import useIsOwner from "@Hooks/react-query/validate/useIsOwner";
 
 import { GetCommentsItem } from "@til-log.lab/tilog-api";
-import { useState } from "react";
 
-interface ParentCommentsProps {
+interface CommentRenderProps {
   comment: GetCommentsItem;
 }
-const CommentChildren = ({ comment }: ParentCommentsProps) => {
+const Comment = ({ comment }: CommentRenderProps) => {
   const [isUpdate, setIsUpdate] = useState(false);
   const handleUpdate = () => {
     setIsUpdate(!isUpdate);
   };
   return (
-    <>
+    <div key={comment.id}>
       <div className="flex items-center gap-x-2 shrink-0">
         <LinkTo href={`/blog/${comment.user.username}`}>
           <UserProfile className="w-10 h-10" avatar={comment.user.avatar}>
@@ -42,7 +42,7 @@ const CommentChildren = ({ comment }: ParentCommentsProps) => {
           <CommentContents comment={comment} />
         )}
       </div>
-    </>
+    </div>
   );
 };
-export default CommentChildren;
+export default Comment;
